requirements.txt
flask==2.0.1
tensorflow-cpu==2.6.0
pandas==1.3.0
numpy==1.21.0
matplotlib==3.4.2
gunicorn==20.1.0